Fix fish shell issues
The script was assuming that if it was being run in the fish shell, the
user would already have created the `~/.config/fish/config.fish` file,
which is not an unrealistic assumption to make. However, it's better
to be safe and follow what I already do for bash and zsh and create
the file if it doesn't exist.

On my spare latop that I use for testing scripts, Homebrew complains
that `PATH` does not include `/usr/local/sbin` because some formulae
installed some executables there. However, I don't see anything in the
`sbin` directory except for a `.keepme` file. I don't see this file
on my main laptop, so it sounds like Homebrew started adding it at
some point this year. If I remove the `.keepme` file, `brew doctor`
no longer complains. To solve this, Homebrew recommends adding
`set -g fish_user_paths "/usr/local/sbin" $fish_user_paths` to
the fish config file, which I don't think hurts.

Finally, for chruby to work in fish, the files that need to be sourced
are the ones from `chruby-fish`, which are in the same location as the
ones from `chruby`, except they end in `.fish` instead of `.sh`.